pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = ""
authors = ["Pedro Contreras Avendano <pedro.contreras.avendano@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.2"
requests = "^2.31.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
beautifulsoup4 = "^4.12.3"
requests-html = "^0.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
